# Make max_transfer_size parameter configurable to optimise mnesia table loading time