--- a/lib/mnesia/src/mnesia.erl
+++ b/lib/mnesia/src/mnesia.erl
@@ -2553,6 +2553,7 @@ system_info2(core_dir) ->  mnesia_monitor:get_env(core_dir);
 system_info2(no_table_loaders) ->  mnesia_monitor:get_env(no_table_loaders);
 system_info2(dc_dump_limit) ->  mnesia_monitor:get_env(dc_dump_limit);
 system_info2(send_compressed) -> mnesia_monitor:get_env(send_compressed);
+system_info2(max_transfer_size) -> mnesia_monitor:get_env(max_transfer_size);
 
 system_info2(Item) -> exit({badarg, Item}).
 
@@ -2598,6 +2599,7 @@ system_info_items(yes) ->
      no_table_loaders,
      dc_dump_limit,
      send_compressed,
+     max_transfer_size,
      version
     ];
 system_info_items(no) ->