--- a/lib/mnesia/src/mnesia_monitor.erl
+++ b/lib/mnesia/src/mnesia_monitor.erl
@@ -691,6 +691,7 @@ env() ->
      no_table_loaders,
      dc_dump_limit,
      send_compressed,
+     max_transfer_size,
      schema
     ].
 
@@ -741,6 +742,8 @@ default_env(dc_dump_limit) ->
     4;
 default_env(send_compressed) ->
     0;
+default_env(max_transfer_size) ->
+    64000;
 default_env(schema) ->
     [].
 
@@ -790,6 +793,7 @@ do_check_type(pid_sort_order, _) -> false;
 do_check_type(no_table_loaders, N) when is_integer(N), N > 0 -> N;
 do_check_type(dc_dump_limit,N) when is_number(N), N > 0 -> N;
 do_check_type(send_compressed, L) when is_integer(L), L >= 0, L =< 9 -> L;
+do_check_type(max_transfer_size, N) when is_integer(N), N > 0 -> N;
 do_check_type(schema, L) when is_list(L) -> L.
 
 bool(true) -> true;